bbx-console/backend/requirements.txt
# BBX Console Backend Dependencies

# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# WebSocket
websockets>=12.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
alembic>=1.13.0

# HTTP Client
httpx>=0.26.0

# Utils
python-multipart>=0.0.6
pyyaml>=6.0
python-dotenv>=1.0.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Type checking
mypy>=1.8.0
